Refactor version manager logic to separate file
